+++ src/api/amount.rs
@@ -5,11 +5,11 @@ use crate::{
 use rosetta::models::{Amount, Currency};
 use std::ops::Deref;
 
-pub fn amount_from_uccd(v: i64) -> Amount {
+pub fn amount_from_uccd(v: i128) -> Amount {
     Amount::new(v.to_string(), Currency::new("CCD".to_string(), 6))
 }
 
-pub fn uccd_from_amount(v: &Amount) -> ApiResult<i64> {
+pub fn uccd_from_amount(v: &Amount) -> ApiResult<i128> {
     validate_currency(v.currency.deref())?;
     v.value.parse().map_err(|_| ApiError::InvalidAmount(v.value.clone()))
 }